Move `cli/local_v1.py` under `utils`
