+++ tests/integration/test_sync_command.py
@@ -8,7 +8,7 @@ from unittest.mock import Mock
 import pytest
 from typing_extensions import ParamSpec
 
-from milatools.cli.local_v1 import LocalV1
+from milatools.utils.local_v1 import LocalV1
 from milatools.utils.remote_v2 import RemoteV2
 from milatools.utils.vscode_utils import (
     extensions_to_install,
